refactor: update namespace paths for enhanced consistency and structure
**Detailed Description:**

1. **Namespace Adjustments:**
   - Refactored namespaces for multiple exception classes:
     - `WrongReturnTypeException`, `SizeException`, and `EmptyDataStructureException` updated from `Exception\DataStructure` to `Exception\Support\DataStructure`.
     - Aligns with the logical organization under a more appropriate "Support" directory for better clarity and maintainability.
   - Updated `InvalidValueObjectError` namespace from `Error\LowLevel` to `Error\ValueObject`, improving hierarchical alignment within the error structure.

2. **Documentation Updates:**
   - Adjusted related comments and annotations to reflect the new namespace paths.
   - Ensures consistency across the codebase for easier navigation and understanding.

3. **Improved Exception References:**
   - Modified all references to the relocated `InvalidValueObjectError` for better contextual accuracy:
     - Updated an instance in `Throwable\ValueObject\Firehub.Code` to reflect the correct path.

**Key Benefits:**
- Enhances overall code modularity and organizational structure by aligning namespaces with intended functionality.
- Improves long-term maintainability and ease of understanding for developers interacting with exception and error classes.
